+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Timeline Full Widget
 
-## 1.0.1 (October 14, 2025)
+## 1.0.1 (October 3, 2025)
 
-* Minor design tweaks
+* Added sticky text settings on the opposite side of the timeline.
+* Corrected animation for markers
 
-## 1.0.0 (October 3, 2025)
+## 1.0.0 (October 1, 2025)
 
 * Minor text changes
 * Add screenshot